+++ java/src/main/java/linalg/algorithms/QRDecomposition.java
@@ -1,0 +1,9 @@
+package linalg.algorithms;
+
+import linalg.matrix.Matrix;
+
+public interface QRDecomposition {
+    Matrix Q();
+    Matrix R();
+    void Decompose();
+}
